UNIS now have 1 byte in config for servo position
one position one byte
all servos and all positions is 12 bytes long

+++ src/modules/unis.h
@@ -14,7 +14,7 @@ struct MtbUnisConfig {
   std::array<uint8_t, UNIS_OUT_CNT> outputsSafe = {0, };
   std::array<size_t, UNIS_IN_CNT> inputsDelay = {0, }; // 0 = 0s, 1 = 0.1s, 15 = 1.5s, min=0, max=15
   uint8_t servoEnabledMask = 0;
-  std::array<uint16_t, UNIS_SERVO_OUT_CNT> servoPosition = {0, };
+  std::array<uint8_t, UNIS_SERVO_OUT_CNT> servoPosition = {0, };
   std::array<uint8_t, UNIS_SERVO_CNT> servoSpeed = {0, };
 
   MtbUnisConfig(const QJsonObject& json) { this->fromJson(json); }
